[#11] Fix test failures (bar one righteous one)
