+++ neoclient/decorators/__init__.py
@@ -1,5 +1,6 @@
 from .client import base_url, verify
 from .common import (
+    accept,
     cookie,
     cookies,
     header,
@@ -7,8 +8,9 @@ from .common import (
     middleware,
     query,
     query_params,
+    referer,
     timeout,
+    user_agent,
 )
-from .headers import accept, referer, user_agent
 from .operation import content, data, files, json, mount, path, path_params
 from .service import service
